# openapi.yml
openapi: "3.0.0"
info:
  version: v0
  title: youyuan core Service
  license:
    name: MIT
servers:
  - url: https://api.youyuan.36node.com{basePath}
    description: The production API server
    variables:
      basePath:
        default: /core/v0
paths:
...
  /repairs:
    post:
      summary: Create a repair order
      operationId: createRepair
      tags:
        - repair
      requestBody:
        description: repair order to add
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairCreateDoc"
      responses:
        "201":
          description: The repair order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repair"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all repair orders
      operationId: listRepairs
      tags:
        - repair
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: id_like
          in: query
          description: repair order id, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: status
          in: query
          description: repair order status
          required: false
          schema:
            type: string
            enum: ["CREATING", "FILLING", "DELIVERED", "REPARING", "REPARED", "PAUSING", "PENDING", "PASSED"]
        - name: createAt_gt
          in: query
          description: greater than createAt
          required: false
          schema:
            type: string
            format: date-time
        - name: createAt_lt
          in: query
          description: less than createAt
          required: false
          schema:
            type: string
            format: date-time
        - name: dispatchWorker_like
          in: query
          description: assignee id, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: plate_like
          in: query
          description: plate, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: crm_like
          in: query
          description: crm, regix
          required: false
          schema:
            type: string
            minLength: 1
        - name: vin_like
          in: query
          description: vin, regix
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: A paged array of repair orders
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repair"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /repairs/{repairId}:
    get:
      summary: Find repair order by id
      operationId: getRepair
      tags:
        - repair
      parameters:
        - name: repairId
          in: path
          required: true
          description: The id of the repair order to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repair"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: update repair order
      operationId: updateRepair
      tags:
        - repair
      parameters:
        - name: repairId
          in: path
          required: true
          description: The id of repair order
          schema:
            type: string
      requestBody:
        description: repair order to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairDoc"
      responses:
        "200":
          description: The repair order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repair"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      description: deletes a single repair order based on the ID supplied
      operationId: deleteRepair
      tags:
        - repair
      parameters:
        - name: repairId
          in: path
          description: ID of repair order to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: repair order deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /repairs/{repairId}/record:
    put:
      summary: update repair order record
      operationId: updateRecord
      tags:
        - repair
      parameters:
        - name: repairId
          in: path
          required: true
          description: The id of repair order
          schema:
            type: string
      requestBody:
        description: repair order record to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairRecord"
      responses:
        "200":
          description: The repair order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /records:
    post:
      summary: Create a record
      operationId: createRecord
      tags:
        - record
      requestBody:
        description: record to add
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordDoc"
      responses:
        "201":
          description: The repair order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      summary: List all records
      operationId: listRecords
      tags:
        - record
      parameters:
        - name: _limit
          in: query
          description: How many items to return at one time (max 100)
          required: false
          schema:
            type: integer
            format: int32
            default: 10
            maximum: 1000
        - name: _offset
          in: query
          description: How many items to escape
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: _sort
          in: query
          description: Fields to sort
          required: false
          schema:
            type: string
        - name: _select
          in: query
          description: Fields to select
          required: false
          schema:
            type: array
            items:
              type: string
        - name: ticket_like
          in: query
          description: repair order id, regix
          required: false
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: A paged array of repair orders
          headers:
            X-Total-Count:
              description: Total count of all warnings match the query
              required: true
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Record"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /records/{recordId}:
    get:
      summary: Find record by id
      operationId: getRecord
      tags:
        - record
      parameters:
        - name: recordId
          in: path
          required: true
          description: The id of the record to retrieve
          schema:
            type: string
      responses:
        "200":
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      summary: Update record
      operationId: updateRecord
      tags:
        - record
      parameters:
        - name: recordId
          in: path
          required: true
          description: The id of record
          schema:
            type: string
      requestBody:
        description: record to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordDoc"
      responses:
        "200":
          description: The record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Record"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      description: deletes a single record based on the ID supplied
      operationId: deleteRecord
      tags:
        - record
      parameters:
        - name: recordId
          in: path
          description: ID of record to delete
          required: true
          schema:
            type: string
      responses:
        "204":
          description: record deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
security:
  - bearerAuth: []
components:
  responses:
...
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
    InternalError:
      description: unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Err"
  schemas:
...
    BackupDoc:
      type: object
      properties:
        itemNo:
          type: string
          description: 物料号
        unitPrice:
          type: string
          description: 单价
        quantity:
          type: number
          description: 数量
        free:
          type: boolean
          description: 是否收费
    WorkerDoc:
      type: object
      properties:
        user:
          type: string
          description: 关联 user id
        mileages:
          type: string
          description: 出工里程
        Workload:
          type: number
          description: 工作量
    RepairCreateDoc:
      type: object
      description: 维修单创建信息
      required:
        - vehicle
        - plate
        - vin
        - address
        - assignee
        - assignAt
        - createBy
      properties:
        maintain:
          type: string
        vehicle:
          type: string
        plate:
          type: string
        vin:
          type: string
        address:
          type: string
        reporter:
          type: string
        reporterPhone:
          type: string
        problem:
          type: string
        assignee:
          type: string
          description: 派工人员 id
        assignAt:
          type: string
          format: date-time
          description: 派送时间
        createBy:
          type: string
          description: 创建者
        ticket:
          type: string
        status:
          type: string
          enum: ["CREATING", "REPARING", "PAUSING", "REJECTED", "PENDING", "PASSED"]
    RepairRecord:
      type: object
      description: 维修单维修记录
      required:
        - signAddress
        - arrivalAt
        - mileages
        - faultCell
        - workHours
        - failReason
        - processMethod
        - faultAttr
        - faultMode
        - livePic
        - resultDesc
        - softwarePic
        - frontPic
        - nameplatePic
        - meterPic
        - batteryPic
        - workers
      properties:
        signAddress:
          type: string
        arrivalAt:
          type: string
          format: date-time
        mileages:
          type: string
        routeMap:
          type: string
        faultCell:
          type: string
        workHours:
          type: string
        failReason:
          type: string
        processMethod:
          type: string
        faultAttr:
          type: string
        faultMode:
          type: string
        backups:
          type: array
          items:
            $ref: "#/components/schemas/BackupDoc"
        livePic:
          type: array
          items:
            type: string
        resultDesc:
          type: string
        softwarePic:
          type: array
          items:
            type: string
        frontPic:
          type: array
          items:
            type: string
        nameplatePic:
          type: array
          items:
            type: string
        meterPic:
          type: array
          items:
            type: string
        batteryPic:
          type: array
          items:
            type: string
        otherPic:
          type: array
          items:
            type: string
        dataFile:
          type: string
        otherCost:
          type: string
        remark:
          type: string
        workers:
          type: array
          items:
            $ref: "#/components/schemas/WorkerDoc"
        pauseReason:
          type: string
        status:
          type: string
          enum: ["CREATING", "REPARING", "PAUSING", "REJECTED", "PENDING", "PASSED"]
    RepairDoc:
      type: object
      properties:
        ticket:
          type: string
          description: 关联的工单 id
        status:
          type: string
          enum: ["CREATING", "REPARING", "PAUSING", "REJECTED", "PENDING", "PASSED"]
          description: 维修单工单状态
        rejected:
          type: boolean
          description: 工单曾经被驳回
        crm:
          type: string
          description: CRM 订单号
        vehicle:
          type: string
          description: 关联车辆 id
        plate:
          type: string
          description: 关联车辆车牌
        vin:
          type: string
          description: 关联车辆vin码
        address:
          type: string
          description: 故障地点
        reporter:
          type: string
          description: 报修人姓名
        reporterPhone:
          type: string
          description: 报修人联系方式
        problem:
          type: string
          description: 问题描述
        assignee:
          type: string
          description: 派工人员 id
        assignAt:
          type: string
          format: date-time
          description: 派送时间
        closeAt:
          type: string
          format: date-time
          description: 完工时间
        record:
          $ref: "#/components/schemas/RepairRecord"
        pauseReason:
          type: string
          description: 暂停原因
        maintain:
          type: string
          description: 关联维修单id
        createBy:
          type: string
    Repair:
      allOf:
        - $ref: "#/components/schemas/RepairDoc"
        - $ref: "#/components/schemas/MongoDefault"
...
    RecordDoc:
      type: object
      properties:
        ticket:
          type: string
          description: 关联工单 id
        title:
          type: string
          description: 操作记录简介
        desc:
          type: string
          description: 操作记录详情
    Record:
      allOf:
        - $ref: "#/components/schemas/RecordDoc"
        - $ref: "#/components/schemas/MongoDefault"
    MongoDefault:
      type: object
      required:
        - id
      properties:
        id:
          type: string
        updateAt:
          type: string
          format: date-time
        updateBy:
          type: string
        createAt:
          type: string
          format: date-time
        createBy:
          type: string
    Err:
      type: object
      required:
        - name
        - message
      properties:
        code:
          type: string
        type:
          type: string
        message:
          type: boolean
        name:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              keyword:
                type: string
              message:
                type: string
              path:
                type: string
              value:
                type: string
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT